# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ANPE CENTRE NORD API",
        "description": "Bienvenue dans la documentation de l'API de l'ANPE CENTRE NORD.<br>Cette API fournit un accès sécurisé aux fonctionnalités du backend, <br>permettant la gestion efficace des ressources liées à l'emploi, aux utilisateurs et aux données associées.",
        "contact": {
            "name": "SYMBOL SARL",
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur test de l'API"
        }
    ],
    "paths": {
...
        "/api/recruteur-offres": {
            "get": {
                "tags": [
                    "Recruteur"
                ],
                "summary": "Liste des offres",
                "description": "Retourne la liste des offres",
                "operationId": "7a67629d19ba5a66aaa93e8332470bff",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recruteur-offres/{slug}": {
            "get": {
                "tags": [
                    "Recruteur"
                ],
                "summary": "Récupération d'une offre par son slug",
                "description": "Retourne une offre",
                "operationId": "2bf6e768188c8b57cbf89a307a412f5e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de l'offre du recruteur à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recruteur-offres-candidatures": {
            "get": {
                "tags": [
                    "Recruteur"
                ],
                "summary": "Liste des candidatures du recruteur",
                "description": "Retourne la liste des candidatures du recruteur",
                "operationId": "d404fd79cf30c051a7676b01bfff8f13",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}